services/job_scraper.py
"""
JobScraperService handles scraping job listings from various job boards

This service focuses on AWS cloud-related job postings from multiple sources.
"""

import requests
from bs4 import BeautifulSoup
import json
import random
from datetime import datetime, timedelta

class JobScraperService:
    """Service for scraping AWS job listings from various platforms"""
    
    def __init__(self):
        self.sources = {
            'linkedin': 'https://www.linkedin.com/jobs/search/?keywords=aws',
            'indeed': 'https://www.indeed.com/jobs?q=aws',
            'glassdoor': 'https://www.glassdoor.com/Job/aws-jobs-SRCH_KO0,3.htm',
            'shine': 'https://www.shine.com/job-search/aws-jobs',
            'internshala': 'https://internshala.com/internships/aws-internship'
        }
    
    def get_source_name_from_url(self, url):
        """
        Extracts the name of the job source from its URL
        
        Args:
            url: The URL of the job source
            
        Returns:
            The name of the job source
        """
        for source_name, source_url in self.sources.items():
            if source_name in url.lower():
                return source_name
        return "other"
    
    def scrape_jobs(self, source):
        """
        Scrapes job listings from a given job source
        
        Args:
            source: The JobSource object to scrape
            
        Returns:
            A list of jobs scraped from the source
        """
        # This would implement actual web scraping logic in a production environment
        # For this demo, we'll return an empty list as jobs would be created manually
        return []