pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "routes>=2.5.1",
    "selenium>=4.31.0",
    "webdriver-manager>=4.0.2",
    "werkzeug>=3.1.3",
]
